refactor Tamiyo, Field Researcher
